refactor: extract reusable bind core and adapter contract
